test: split board persistence process and polling helpers
