+++ tests/functional/sessions/restart/board_persistence_cli_test.go
@@ -4,362 +4,11 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-	"net"
-	"net/http"
-	"os"
 	"os/exec"
-	"path/filepath"
-	"runtime"
-	"strings"
-	"sync"
 	"testing"
-	"time"
 
 	"github.com/portpowered/infinite-you/internal/builtcliacceptance"
-	"github.com/portpowered/infinite-you/internal/testutil"
-	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
-	"github.com/portpowered/infinite-you/tests/functional/internal/support"
 )
-
-type boardPersistenceDaemon struct {
-	cmd        *exec.Cmd
-	baseURL    string
-	sessionID  string
-	factoryDir string
-	homeDir    string
-	logDir     string
-	recordPath string
-	stdout     *bytes.Buffer
-	stderr     *bytes.Buffer
-	done       chan struct{}
-	mu         sync.Mutex
-	waitErr    error
-	stopped    bool
-}
-
-func buildBoardPersistenceBinary(t *testing.T) string {
-	t.Helper()
-	binaryName := "you"
-	if runtime.GOOS == "windows" {
-		binaryName += ".exe"
-	}
-	binaryPath := filepath.Join(t.TempDir(), binaryName)
-	build := exec.CommandContext(t.Context(), "go", "build", "-o", binaryPath, "./cmd/factory")
-	build.Dir = testutil.MustRepoRoot(t)
-	output, err := build.CombinedOutput()
-	if err != nil {
-		t.Fatalf("build fresh you binary: %v\n%s", err, output)
-	}
-	return binaryPath
-}
-
-func startBoardPersistenceDaemon(
-	t *testing.T,
-	binaryPath, factoryDir, homeDir, recordPath, releasePath string,
-) *boardPersistenceDaemon {
-	t.Helper()
-	daemon := startBoardPersistenceDaemonProcess(t, binaryPath, factoryDir, homeDir, recordPath, releasePath)
-	waitForBoardDaemonReady(t, daemon, 45*time.Second)
-	daemon.sessionID = waitForBoardSessionID(t, daemon.baseURL, 30*time.Second)
-	t.Logf("isolated daemon live session ID: %q", daemon.sessionID)
-	return daemon
-}
-
-func startBoardPersistenceDaemonProcess(
-	t *testing.T,
-	binaryPath, factoryDir, homeDir, recordPath, releasePath string,
-) *boardPersistenceDaemon {
-	t.Helper()
-	address := reserveBoardPersistenceAddress(t)
-	command := exec.CommandContext(
-		t.Context(),
-		binaryPath,
-		"run", "--dir", factoryDir,
-		"--continuously", "--with-server",
-		"--listen", address,
-		"--record", recordPath,
-	)
-	command.Dir = factoryDir
-	command.Env = append(
-		builtcliacceptance.ProcessEnvForIsolatedHome(homeDir),
-		boardPersistenceHelperEnv+"="+boardPersistenceHelperEnvValue,
-		boardPersistenceReleaseEnv+"="+releasePath,
-	)
-	configureBoardPersistenceCommand(command)
-	var stdout, stderr bytes.Buffer
-	command.Stdout = &stdout
-	command.Stderr = &stderr
-	daemon := &boardPersistenceDaemon{
-		cmd:        command,
-		baseURL:    "http://" + address,
-		factoryDir: factoryDir,
-		homeDir:    homeDir,
-		recordPath: recordPath,
-		stdout:     &stdout,
-		stderr:     &stderr,
-		done:       make(chan struct{}),
-		logDir:     filepath.Join(homeDir, ".you-agent-factory", "logs"),
-	}
-	if err := command.Start(); err != nil {
-		t.Fatalf("start isolated you daemon: %v", err)
-	}
-	go func() {
-		err := command.Wait()
-		daemon.mu.Lock()
-		daemon.waitErr = err
-		daemon.mu.Unlock()
-		close(daemon.done)
-	}()
-	t.Cleanup(daemon.cleanup)
-	return daemon
-}
-
-func waitForBoardPersistenceDaemonExit(t *testing.T, daemon *boardPersistenceDaemon, timeout time.Duration) {
-	t.Helper()
-	// A startup failure is observed through the real child-process exit rather
-	// than a fixed sleep; this is the process-boundary behavior under test.
-	timer := time.NewTimer(timeout)
-	defer timer.Stop()
-	select {
-	case <-daemon.done:
-	case <-timer.C:
-		t.Fatalf("corrupt-recording daemon did not exit within %s\nstdout=%s\nstderr=%s", timeout, daemon.stdout.String(), daemon.stderr.String())
-	}
-}
-
-func (daemon *boardPersistenceDaemon) kill(t *testing.T) {
-	t.Helper()
-	if daemon == nil {
-		return
-	}
-	daemon.mu.Lock()
-	if daemon.stopped {
-		daemon.mu.Unlock()
-		return
-	}
-	daemon.mu.Unlock()
-	if err := daemon.cmd.Process.Kill(); err != nil && !errors.Is(err, os.ErrProcessDone) {
-		t.Fatalf("hard-kill isolated you daemon: %v", err)
-	}
-	timer := time.NewTimer(20 * time.Second)
-	defer timer.Stop()
-	select {
-	case <-daemon.done:
-		daemon.mu.Lock()
-		daemon.stopped = true
-		daemon.mu.Unlock()
-	case <-timer.C:
-		t.Fatalf("hard-killed isolated you daemon did not exit within 20s\nstdout=%s\nstderr=%s", daemon.stdout.String(), daemon.stderr.String())
-	}
-}
-
-func (daemon *boardPersistenceDaemon) cleanup() {
-	if daemon == nil {
-		return
-	}
-	daemon.mu.Lock()
-	if daemon.stopped {
-		daemon.mu.Unlock()
-		return
-	}
-	daemon.stopped = true
-	daemon.mu.Unlock()
-	select {
-	case <-daemon.done:
-	default:
-		_ = daemon.cmd.Process.Kill()
-		<-daemon.done
-	}
-}
-
-func (daemon *boardPersistenceDaemon) stop(t *testing.T) {
-	t.Helper()
-	daemon.mu.Lock()
-	if daemon.stopped {
-		daemon.mu.Unlock()
-		return
-	}
-	daemon.mu.Unlock()
-	if err := interruptBoardPersistenceProcess(daemon.cmd); err != nil {
-		t.Fatalf("interrupt isolated you daemon: %v", err)
-	}
-	timer := time.NewTimer(20 * time.Second)
-	defer timer.Stop()
-	select {
-	case <-daemon.done:
-		if err := daemon.waitError(); err != nil {
-			if !boardPersistenceCleanExit(err) {
-				t.Fatalf("isolated you daemon shutdown error = %v\nstdout=%s\nstderr=%s", err, daemon.stdout.String(), daemon.stderr.String())
-			}
-		}
-		daemon.mu.Lock()
-		daemon.stopped = true
-		daemon.mu.Unlock()
-	case <-timer.C:
-		t.Fatalf("isolated you daemon did not stop within 20s\nstdout=%s\nstderr=%s", daemon.stdout.String(), daemon.stderr.String())
-	}
-}
-
-func (daemon *boardPersistenceDaemon) waitError() error {
-	daemon.mu.Lock()
-	defer daemon.mu.Unlock()
-	return daemon.waitErr
-}
-
-func waitForBoardDaemonReady(t *testing.T, daemon *boardPersistenceDaemon, timeout time.Duration) {
-	t.Helper()
-	deadline := time.NewTimer(timeout)
-	defer deadline.Stop()
-	// The real child process exposes no parent readiness channel. This bounded
-	// public /status observation is the unavoidable process-bound synchronization
-	// for the isolated-daemon acceptance test.
-	ticker := time.NewTicker(25 * time.Millisecond)
-	defer ticker.Stop()
-	client := &http.Client{Timeout: 2 * time.Second}
-	for {
-		select {
-		case <-daemon.done:
-			dumpBoardPersistenceDiagnostics(t, daemon)
-			t.Fatalf("isolated you daemon exited before readiness: %v\nstdout=%s\nstderr=%s", daemon.waitError(), daemon.stdout.String(), daemon.stderr.String())
-		case <-ticker.C:
-			request, err := http.NewRequestWithContext(t.Context(), http.MethodGet, daemon.baseURL+"/status", nil)
-			if err != nil {
-				t.Fatalf("build daemon readiness request: %v", err)
-			}
-			response, err := client.Do(request)
-			if err != nil {
-				continue
-			}
-			var status factoryapi.StatusResponse
-			decodeErr := json.NewDecoder(response.Body).Decode(&status)
-			_ = response.Body.Close()
-			if response.StatusCode == http.StatusOK && decodeErr == nil && status.RuntimeStatus != "" {
-				return
-			}
-		case <-deadline.C:
-			t.Fatalf("timed out waiting for isolated you daemon readiness at %s\nstdout=%s\nstderr=%s", daemon.baseURL, daemon.stdout.String(), daemon.stderr.String())
-		}
-	}
-}
-
-func dumpBoardPersistenceDiagnostics(t *testing.T, daemon *boardPersistenceDaemon) {
-	t.Helper()
-	t.Logf("daemon diagnostic paths: factory=%q home=%q record=%q", daemon.factoryDir, daemon.homeDir, daemon.recordPath)
-	logsRoot := filepath.Join(daemon.homeDir, ".you-agent-factory", "logs")
-	_ = filepath.WalkDir(logsRoot, func(path string, entry os.DirEntry, err error) error {
-		if err != nil || entry.IsDir() || !strings.HasSuffix(strings.ToLower(entry.Name()), ".log") {
-			return nil
-		}
-		contents, readErr := os.ReadFile(path)
-		if readErr != nil {
-			return nil
-		}
-		if len(contents) > 8192 {
-			contents = contents[len(contents)-8192:]
-		}
-		t.Logf("daemon runtime log tail %q (%d bytes): %s", path, len(contents), contents)
-		return nil
-	})
-}
-
-func waitForBoardPersistenceSnapshot(t *testing.T, path, wantSessionID string, timeout time.Duration) []byte {
-	t.Helper()
-	// The durable snapshot is committed by the isolated daemon child, and the
-	// parent has no synchronization channel for that filesystem write. Polling
-	// the file is the only deterministic observation of the commit boundary;
-	// the bounded timeout turns a failed child write into a useful test failure.
-	deadline := time.NewTimer(timeout)
-	defer deadline.Stop()
-	ticker := time.NewTicker(25 * time.Millisecond)
-	defer ticker.Stop()
-	var lastErr error
-	for {
-		contents, err := os.ReadFile(path)
-		if err == nil && len(contents) > 0 {
-			var snapshot struct {
-				Session struct {
-					SessionID string `json:"sessionId"`
-				} `json:"session"`
-			}
-			if err := json.Unmarshal(contents, &snapshot); err == nil && snapshot.Session.SessionID == wantSessionID {
-				return contents
-			}
-			lastErr = fmt.Errorf("durable snapshot was empty or session identity was not %q", wantSessionID)
-		} else {
-			lastErr = err
-		}
-		select {
-		case <-ticker.C:
-		case <-deadline.C:
-			t.Fatalf("timed out waiting for valid durable snapshot %q: %v", path, lastErr)
-		}
-	}
-}
-
-func waitForBoardPersistenceLogMessage(
-	t *testing.T,
-	daemon *boardPersistenceDaemon,
-	fragments []string,
-	timeout time.Duration,
-) {
-	t.Helper()
-	// The recovery warning is appended by the isolated child process after boot,
-	// with no test-owned logging edge back to the parent. Polling the runtime log
-	// is therefore the required process-boundary observation; the timeout keeps
-	// a missing warning actionable without using a fixed sleep.
-	deadline := time.NewTimer(timeout)
-	defer deadline.Stop()
-	ticker := time.NewTicker(25 * time.Millisecond)
-	defer ticker.Stop()
-	for {
-		found := false
-		_ = filepath.WalkDir(daemon.logDir, func(path string, entry os.DirEntry, err error) error {
-			if err != nil || entry.IsDir() || !strings.HasSuffix(strings.ToLower(entry.Name()), ".log") {
-				return nil
-			}
-			contents, readErr := os.ReadFile(path)
-			if readErr != nil {
-				return nil
-			}
-			for _, fragment := range fragments {
-				if !strings.Contains(string(contents), fragment) {
-					return nil
-				}
-			}
-			found = true
-			return filepath.SkipAll
-		})
-		if found {
-			return
-		}
-		select {
-		case <-ticker.C:
-		case <-deadline.C:
-			t.Fatalf("timed out waiting for recovery warning in runtime logs under %q", daemon.logDir)
-		}
-	}
-}
-
-func reserveBoardPersistenceAddress(t *testing.T) string {
-	t.Helper()
-	for {
-		listener, err := net.Listen("tcp4", "127.0.0.1:0")
-		if err != nil {
-			t.Fatalf("reserve isolated daemon loopback port: %v", err)
-		}
-		address := listener.Addr().String()
-		port := listener.Addr().(*net.TCPAddr).Port
-		if err := listener.Close(); err != nil {
-			t.Fatalf("release isolated daemon loopback port: %v", err)
-		}
-		if port != 7437 {
-			return address
-		}
-	}
-}
 
 func runBoardPersistenceCLI(
 	ctx context.Context,
@@ -394,125 +43,5 @@ func submitBatchThroughCLI(
 	}
 	if acknowledgement.RequestID != requestID || acknowledgement.WorkCount != wantWorkCount {
 		t.Fatalf("you submit batch acknowledgement = %#v, want requestId %q and workCount %d", acknowledgement, requestID, wantWorkCount)
-	}
-}
-
-func waitForBoardStates(t *testing.T, baseURL string, want map[string]string, timeout time.Duration) factoryapi.ListWorkResponse {
-	t.Helper()
-	deadline := time.NewTimer(timeout)
-	defer deadline.Stop()
-	// Work has no process-parent notification, so state convergence is observed
-	// through the public session list with a bounded ticker rather than a fixed
-	// sleep. This is the process-bound wait required by this test only.
-	ticker := time.NewTicker(25 * time.Millisecond)
-	defer ticker.Stop()
-	var last factoryapi.ListWorkResponse
-	var lastErr error
-	for {
-		listed, err := readBoardWorkList(t.Context(), baseURL)
-		if err == nil {
-			last = listed
-			if boardStatesMatch(listed, want) {
-				return listed
-			}
-		} else {
-			lastErr = err
-		}
-		select {
-		case <-ticker.C:
-		case <-deadline.C:
-			t.Fatalf("timed out waiting for Work states %#v; last list=%#v, last error=%v", want, last.Results, lastErr)
-		}
-	}
-}
-
-func readBoardWorkList(ctx context.Context, baseURL string) (factoryapi.ListWorkResponse, error) {
-	request, err := http.NewRequestWithContext(ctx, http.MethodGet, support.DefaultSessionWorkURL(baseURL, "/work"), nil)
-	if err != nil {
-		return factoryapi.ListWorkResponse{}, err
-	}
-	response, err := http.DefaultClient.Do(request)
-	if err != nil {
-		return factoryapi.ListWorkResponse{}, err
-	}
-	defer response.Body.Close()
-	if response.StatusCode != http.StatusOK {
-		body, _ := io.ReadAll(response.Body)
-		return factoryapi.ListWorkResponse{}, fmt.Errorf("GET /work status %d: %s", response.StatusCode, strings.TrimSpace(string(body)))
-	}
-	var listed factoryapi.ListWorkResponse
-	if err := json.NewDecoder(response.Body).Decode(&listed); err != nil {
-		return factoryapi.ListWorkResponse{}, err
-	}
-	return listed, nil
-}
-
-func boardStatesMatch(listed factoryapi.ListWorkResponse, want map[string]string) bool {
-	if len(listed.Results) != len(want) {
-		return false
-	}
-	seen := make(map[string]struct{}, len(listed.Results))
-	for _, item := range listed.Results {
-		workID := support.StringPointerValue(item.WorkId)
-		state := ""
-		if item.State != nil {
-			state = item.State.Name
-		}
-		if _, duplicate := seen[workID]; duplicate || want[workID] != state {
-			return false
-		}
-		seen[workID] = struct{}{}
-	}
-	return len(seen) == len(want)
-}
-
-func assertBoardList(t *testing.T, listed factoryapi.ListWorkResponse, expected map[string]boardPersistenceExpectedWork) {
-	t.Helper()
-	if len(listed.Results) != len(expected) {
-		t.Fatalf("Work list count = %d, want %d: %#v", len(listed.Results), len(expected), listed.Results)
-	}
-	seen := make(map[string]struct{}, len(listed.Results))
-	for _, item := range listed.Results {
-		workID := support.StringPointerValue(item.WorkId)
-		if _, duplicate := seen[workID]; duplicate {
-			t.Fatalf("Work list duplicated Work ID %q", workID)
-		}
-		want, ok := expected[workID]
-		if !ok {
-			t.Fatalf("Work list returned unexpected Work ID %q: %#v", workID, item)
-		}
-		seen[workID] = struct{}{}
-		assertBoardWork(t, item, want)
-	}
-}
-
-func assertBoardCLIListAndShows(
-	t *testing.T,
-	ctx context.Context,
-	daemon *boardPersistenceDaemon,
-	binaryPath, factoryDir, homeDir string,
-	expected map[string]boardPersistenceExpectedWork,
-) {
-	t.Helper()
-	listOutput, err := runBoardPersistenceCLI(ctx, binaryPath, factoryDir, homeDir, daemon.baseURL, "--json", "work", "list")
-	if err != nil {
-		t.Fatalf("you work list: %v\noutput:\n%s", err, listOutput)
-	}
-	var listed factoryapi.ListWorkResponse
-	if err := json.Unmarshal(bytes.TrimSpace(listOutput), &listed); err != nil {
-		t.Fatalf("decode you work list: %v\noutput:\n%s", err, listOutput)
-	}
-	assertBoardList(t, listed, expected)
-
-	for workID, want := range expected {
-		showOutput, err := runBoardPersistenceCLI(ctx, binaryPath, factoryDir, homeDir, daemon.baseURL, "--json", "work", "show", workID)
-		if err != nil {
-			t.Fatalf("you work show %s: %v\noutput:\n%s", workID, err, showOutput)
-		}
-		var shown factoryapi.Work
-		if err := json.Unmarshal(bytes.TrimSpace(showOutput), &shown); err != nil {
-			t.Fatalf("decode you work show %s: %v\noutput:\n%s", workID, err, showOutput)
-		}
-		assertBoardWork(t, shown, want)
 	}
 }
